[fix] rename "Unit" to "Integration"
