+++ backend-app/tests/Integration/Web/ContactTest.php
@@ -1,13 +1,15 @@
 <?php
-namespace Tests\Unit\Web;
+namespace Tests\Integration\Web;
 
 use TestCase;
 use Illuminate\Foundation\Testing\WithoutMiddleware;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
-class HomeTest extends TestCase
+class ContactTest extends TestCase
 {
+    use DatabaseMigrations;
+
     public function testIndex()
     {
         $response = $this->get(route('web.root.index'));
