+++ backend-app/tests/Integration/Web/HomeTest.php
@@ -1,15 +1,13 @@
 <?php
-namespace Tests\Unit\Web;
+namespace Tests\Integration\Web;
 
 use TestCase;
 use Illuminate\Foundation\Testing\WithoutMiddleware;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
-class ContactTest extends TestCase
+class HomeTest extends TestCase
 {
-    use DatabaseMigrations;
-
     public function testIndex()
     {
         $response = $this->get(route('web.root.index'));
