+++ backend-app/tests/Integration/Web/TagTest.php
@@ -1,28 +1,28 @@
 <?php
-namespace Tests\Unit\Web;
+namespace Tests\Integration\Web;
 
 use TestCase;
 use Illuminate\Foundation\Testing\WithoutMiddleware;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
-use App\Models\Category;
+use App\Models\Tag;
 
-class CategoryTest extends TestCase
+class TagTest extends TestCase
 {
     use DatabaseMigrations;
 
     public function testIndex()
     {
-        $response = $this->get(route('web.categories.index'));
+        $response = $this->get(route('web.tags.index'));
 
         $response->assertStatus(200);
     }
 
     public function testGetPosts()
     {
-        $category = Category::first();
+        $tag = Tag::first();
 
-        $response = $this->get(route('web.posts.categories.getPosts', $category->name));
+        $response = $this->get(route('web.posts.tags.getPosts', $tag->name));
 
         $response->assertStatus(200);
     }
